requirements.txt
asgiref==3.4.1
Django==3.2.6
django-filter==2.4.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Markdown==3.3.4
phonenumbers==8.12.30
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
